v3.1 commit
Changed compilation options to P-Code instead of native code.
- This change was made due to version 3.0 being incorrectly detected as @Trojan:Win32/Azden.A!cl@ by Windows Defender in some cases.
Added VITS test signal option
Shows the four most recently used config files in the File menu. Also added an option to Settings to clear this list.
Added the ability to check for updates. This is a manual process and does not run automatically under any circumstance.
Added two options in the Settings menu (for Windows only).
- Allow for special handling of BSB channels. This allows you to compile five separate hacktv.exe files, one for each channel, with a different channel ID in each.
- Temporarily disable some CPU C-states while hacktv is running. This can reduce choppy video on some systems due to timing issues when the CPU enters a low power state
Added an option to remove all settings from the registry (resets all options to defaults)
Cosmetic change to the way the arcorrection value in the config file works:
- If the value is not specified, the option in the GUI is unchecked
- If the value is specified, the option in the GUI is checked and set to whatever is specified
Corrected a bug which prevented a config file in the new format from loading if the sample rate was not specified (now defaults to 16 MHz)