Major update, see description
Rewrote much of the internals to read video modes, scrambling systems and bandplans from an external file named Modes.ini. This allows for greater customisation and also to allow new modes to be added without updating the main code. A copy of Modes.ini has been included in the source as an offline fallback but the application will download the latest version from here on each startup. This can be disabled in settings to force using a local copy and slightly improve startup speed.

Improved handling of comments in INI files (including HTV config files)

Split some functions and methods into a separate class file