Added youtube-dl support
Downloading videos using youtube-dl is now supported.
youtube-dl should be in the system path or the same directory as the JAR file.
youtube.com and youtu.be URLs are auto-detected. If you want to use this on another supported site (e.g. Dailymotion or Facebook), prefix the URL with ytdl: and this will force the youtube-dl handler to kick in.